:rotating_light: breaking change; move all compiled assets to /dist
