+++ js/app.js
@@ -1,0 +1,3 @@
+import Reveal from './reveal.js'
+
+window.Reveal = Reveal();
